refactor: support middleware sub-package
