+++ packages/react/src/components/logged-out/index.tsx
@@ -1,4 +1,4 @@
-import { useSlashID } from "../../main";
+import { useSlashID } from "../..";
 
 type Props = {
   children: React.ReactNode;
