+++ packages/react/src/hooks/use-organizations.test.tsx
@@ -1,5 +1,5 @@
 import { act, render, screen, waitFor } from "@testing-library/react";
-import { useOrganizations } from "../main";
+import { useOrganizations } from "..";
 import { faker } from "@faker-js/faker";
 import { OrganizationDetails } from "@slashid/slashid";
 import { TestSlashIDProvider } from "../context/test-slash-id-provider";
